copy from intra repository

+++ signals/signal_handler.c
@@ -1,17 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   signal_handler.c                                   :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: bootjan <bootjan@student.42.fr>            +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/11/27 13:57:09 by tsteur            #+#    #+#             */
-/*   Updated: 2023/12/09 18:47:08 by bootjan          ###   ########.fr       */
+/*                                                        ::::::::            */
+/*   signal_handler.c                                   :+:    :+:            */
+/*                                                     +:+                    */
+/*   By: bootjan <bootjan@student.42.fr>              +#+                     */
+/*                                                   +#+                      */
+/*   Created: 2023/11/27 13:57:09 by tsteur        #+#    #+#                 */
+/*   Updated: 2023/12/15 17:20:50 by bschaafs      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "signal_handler.h"
 #include "minishell.h"
+
+static volatile sig_atomic_t	g_lastsig;
 
 void	sighandler_interactive(int signum)
 {
@@ -19,7 +20,7 @@ void	sighandler_interactive(int signum)
 	{
 		write(STDOUT_FILENO, "\n", 1);
 		rl_on_new_line();
-		// rl_replace_line("", 0);
+		rl_replace_line("", 0);
 		rl_redisplay();
 	}
 	else if (signum == SIGQUIT)
@@ -27,31 +28,41 @@ void	sighandler_interactive(int signum)
 		rl_on_new_line();
 		rl_redisplay();
 	}
+	g_lastsig = signum;
 }
 
 void	sighandler_running(int signum)
 {
-	signal(signum, SIG_IGN);
-	kill(0, signum);
+	if (g_lastsig != signum)
+		kill(0, signum);
+	g_lastsig = signum;
 }
 
-void	sighandler_heredoc(int signum)
+void	sighandler_in_heredoc(int signum)
 {
 	if (signum == SIGINT)
 	{
-		signal(signum, SIG_IGN);
-		kill(0, signum);
+		write(STDOUT_FILENO, "\n", 1);
+		exit(130);
+		rl_on_new_line();
 	}
 	else if (signum == SIGQUIT)
 	{
-		
 	}
+	g_lastsig = signum;
 }
 
 void	set_sighandler(__sighandler_t sighandler)
 {
 	struct sigaction	sa;
 
+	if (sighandler == NULL)
+	{
+		signal(SIGINT, SIG_DFL);
+		signal(SIGQUIT, SIG_DFL);
+		return ;
+	}
+	g_lastsig = 0;
 	sa.sa_flags = 0;
 	sigemptyset(&sa.sa_mask);
 	sa.sa_handler = sighandler;
